+++ node-ev/lib/MessageBuilder.js
@@ -1,0 +1,15 @@
+const messageTypeId = 3
+
+class MessageBuilder {
+
+  static response (uniqueId, payload) {
+    return [
+      messageTypeId,
+      uniqueId,
+      payload
+    ]
+  }
+
+}
+
+module.exports = MessageBuilder